Users now have to confirm that they own an email when they try to add it to their account.
Summary: Scrapped the AddEmailForm on the logged in homepage. Now we use
the ClaimEmailForm there instead. Deleted the AddEmailForm from the
codebase entirely.

Known issue: The groups an email has subscribed to aren't "brought with
the email" when the email is moved to a new account. This bug existed
before this revision.