Tried to making text msg truncation nicer.
Summary: We now split up msgs into chunks marked with (1/N), (2/N), etc.

Known issues: Not clear if it works properly. The SMS that appears when you
access /twilio_reply/ on the website seems correct, but we can't test
the actual tests very well due to the twilio sandbox msg.