Improved logged-in-navbar create/join group buttons
Summary:

Instead of having a link to join a group and a button to create
in the navbar, I've added buttons to do both. This is similar to
the landing page.

Known issues:

To get around issues with redirects and post I had to hardcode
the form submit URLs in create_or_join.html.

If you try to join a group that you're already a member of, you
get a mailman error. Not sure that this has anything to do with
this commit, though. (Bug may have been pre-existing.)